refactor: use a single buffer implementation and await on nested buffers at unroll stage
