+++ packages/server-renderer/__tests__/renderToStream.spec.ts
@@ -11,11 +11,8 @@ import {
   createStaticVNode
 } from 'vue'
 import { escapeHtml, mockWarn } from '@vue/shared'
-import {
-  renderToStream as _renderToStream,
-  renderComponent,
-  renderSlot
-} from '../src/renderToStream'
+import { renderComponent, renderSlot } from '../src/render'
+import { renderToStream as _renderToStream } from '../src/renderToStream'
 import { Readable } from 'stream'
 
 mockWarn()
