+++ packages/server-renderer/src/render.ts
@@ -69,303 +69,314 @@ export type SSRSlot = (
   scopeId: string | null
 ) => void
 
-export function createServerRenderer(createBuffer: () => BufferInstance) {
-  function renderComponent(
-    comp: Component,
-    props: Props | null = null,
-    children: Slots | SSRSlots | null = null,
-    parentComponent: ComponentInternalInstance | null = null
-  ): SSRBuffer | Promise<SSRBuffer> {
-    return renderComponentVNode(
-      createVNode(comp, props, children),
-      parentComponent
-    )
-  }
-
-  function renderComponentVNode(
-    vnode: VNode,
-    parentComponent: ComponentInternalInstance | null = null
-  ): SSRBuffer | Promise<SSRBuffer> {
-    const instance = createComponentInstance(vnode, parentComponent, null)
-    const res = setupComponent(instance, true /* isSSR */)
-    if (isPromise(res)) {
-      return res
-        .catch(err => {
-          warn(`[@vue/server-renderer]: Uncaught error in async setup:\n`, err)
-        })
-        .then(() => renderComponentSubTree(instance))
-    } else {
-      return renderComponentSubTree(instance)
+// Each component has a buffer array.
+// A buffer array can contain one of the following:
+// - plain string
+// - A resolved buffer (recursive arrays of strings that can be unrolled
+//   synchronously)
+// - An async buffer (a Promise that resolves to a resolved buffer)
+function createBuffer(): BufferInstance {
+  let appendable = false
+  const buffer: SSRBuffer = []
+  return {
+    getBuffer(): SSRBuffer {
+      // Return static buffer and await on items during unroll stage
+      return buffer
+    },
+    push(item: SSRBufferItem) {
+      const isStringItem = isString(item)
+      if (appendable && isStringItem) {
+        buffer[buffer.length - 1] += item as string
+      } else {
+        buffer.push(item)
+      }
+      appendable = isStringItem
     }
   }
+}
 
-  function renderComponentSubTree(
-    instance: ComponentInternalInstance
-  ): SSRBuffer | Promise<SSRBuffer> {
-    const comp = instance.type as Component
-    const { getBuffer, push } = createBuffer()
-    if (isFunction(comp)) {
+export function renderComponent(
+  comp: Component,
+  props: Props | null = null,
+  children: Slots | SSRSlots | null = null,
+  parentComponent: ComponentInternalInstance | null = null
+): SSRBuffer | Promise<SSRBuffer> {
+  return renderComponentVNode(
+    createVNode(comp, props, children),
+    parentComponent
+  )
+}
+
+export function renderComponentVNode(
+  vnode: VNode,
+  parentComponent: ComponentInternalInstance | null = null
+): SSRBuffer | Promise<SSRBuffer> {
+  const instance = createComponentInstance(vnode, parentComponent, null)
+  const res = setupComponent(instance, true /* isSSR */)
+  if (isPromise(res)) {
+    return res
+      .catch(err => {
+        warn(`[@vue/server-renderer]: Uncaught error in async setup:\n`, err)
+      })
+      .then(() => renderComponentSubTree(instance))
+  } else {
+    return renderComponentSubTree(instance)
+  }
+}
+
+function renderComponentSubTree(
+  instance: ComponentInternalInstance
+): SSRBuffer | Promise<SSRBuffer> {
+  const comp = instance.type as Component
+  const { getBuffer, push } = createBuffer()
+  if (isFunction(comp)) {
+    renderVNode(push, renderComponentRoot(instance), instance)
+  } else {
+    if (!instance.render && !comp.ssrRender && isString(comp.template)) {
+      comp.ssrRender = ssrCompile(comp.template, instance)
+    }
+
+    if (comp.ssrRender) {
+      // optimized
+      // set current rendering instance for asset resolution
+      setCurrentRenderingInstance(instance)
+      comp.ssrRender(instance.proxy, push, instance)
+      setCurrentRenderingInstance(null)
+    } else if (instance.render) {
       renderVNode(push, renderComponentRoot(instance), instance)
     } else {
-      if (!instance.render && !comp.ssrRender && isString(comp.template)) {
-        comp.ssrRender = ssrCompile(comp.template, instance)
-      }
+      warn(
+        `Component ${
+          comp.name ? `${comp.name} ` : ``
+        } is missing template or render function.`
+      )
+      push(`<!---->`)
+    }
+  }
+  return getBuffer()
+}
 
-      if (comp.ssrRender) {
-        // optimized
-        // set current rendering instance for asset resolution
-        setCurrentRenderingInstance(instance)
-        comp.ssrRender(instance.proxy, push, instance)
-        setCurrentRenderingInstance(null)
-      } else if (instance.render) {
-        renderVNode(push, renderComponentRoot(instance), instance)
+function renderVNode(
+  push: PushFn,
+  vnode: VNode,
+  parentComponent: ComponentInternalInstance
+) {
+  const { type, shapeFlag, children } = vnode
+  switch (type) {
+    case Text:
+      push(escapeHtml(children as string))
+      break
+    case Comment:
+      push(
+        children ? `<!--${escapeHtmlComment(children as string)}-->` : `<!---->`
+      )
+      break
+    case Static:
+      push(children as string)
+      break
+    case Fragment:
+      push(`<!--[-->`) // open
+      renderVNodeChildren(push, children as VNodeArrayChildren, parentComponent)
+      push(`<!--]-->`) // close
+      break
+    default:
+      if (shapeFlag & ShapeFlags.ELEMENT) {
+        renderElementVNode(push, vnode, parentComponent)
+      } else if (shapeFlag & ShapeFlags.COMPONENT) {
+        push(renderComponentVNode(vnode, parentComponent))
+      } else if (shapeFlag & ShapeFlags.TELEPORT) {
+        renderTeleportVNode(push, vnode, parentComponent)
+      } else if (shapeFlag & ShapeFlags.SUSPENSE) {
+        renderVNode(
+          push,
+          normalizeSuspenseChildren(vnode).content,
+          parentComponent
+        )
       } else {
         warn(
-          `Component ${
-            comp.name ? `${comp.name} ` : ``
-          } is missing template or render function.`
+          '[@vue/server-renderer] Invalid VNode type:',
+          type,
+          `(${typeof type})`
         )
-        push(`<!---->`)
       }
-    }
-    return getBuffer()
+  }
+}
+
+function renderVNodeChildren(
+  push: PushFn,
+  children: VNodeArrayChildren,
+  parentComponent: ComponentInternalInstance
+) {
+  for (let i = 0; i < children.length; i++) {
+    renderVNode(push, normalizeVNode(children[i]), parentComponent)
+  }
+}
+
+function renderElementVNode(
+  push: PushFn,
+  vnode: VNode,
+  parentComponent: ComponentInternalInstance
+) {
+  const tag = vnode.type as string
+  let { props, children, shapeFlag, scopeId, dirs } = vnode
+  let openTag = `<${tag}`
+
+  if (dirs) {
+    props = applySSRDirectives(vnode, props, dirs)
   }
 
-  function renderVNode(
-    push: PushFn,
-    vnode: VNode,
-    parentComponent: ComponentInternalInstance
-  ) {
-    const { type, shapeFlag, children } = vnode
-    switch (type) {
-      case Text:
+  if (props) {
+    openTag += ssrRenderAttrs(props, tag)
+  }
+
+  if (scopeId) {
+    openTag += ` ${scopeId}`
+    const treeOwnerId = parentComponent && parentComponent.type.__scopeId
+    // vnode's own scopeId and the current rendering component's scopeId is
+    // different - this is a slot content node.
+    if (treeOwnerId && treeOwnerId !== scopeId) {
+      openTag += ` ${treeOwnerId}-s`
+    }
+  }
+
+  push(openTag + `>`)
+  if (!isVoidTag(tag)) {
+    let hasChildrenOverride = false
+    if (props) {
+      if (props.innerHTML) {
+        hasChildrenOverride = true
+        push(props.innerHTML)
+      } else if (props.textContent) {
+        hasChildrenOverride = true
+        push(escapeHtml(props.textContent))
+      } else if (tag === 'textarea' && props.value) {
+        hasChildrenOverride = true
+        push(escapeHtml(props.value))
+      }
+    }
+    if (!hasChildrenOverride) {
+      if (shapeFlag & ShapeFlags.TEXT_CHILDREN) {
         push(escapeHtml(children as string))
-        break
-      case Comment:
-        push(
-          children
-            ? `<!--${escapeHtmlComment(children as string)}-->`
-            : `<!---->`
-        )
-        break
-      case Static:
-        push(children as string)
-        break
-      case Fragment:
-        push(`<!--[-->`) // open
+      } else if (shapeFlag & ShapeFlags.ARRAY_CHILDREN) {
         renderVNodeChildren(
           push,
           children as VNodeArrayChildren,
           parentComponent
         )
-        push(`<!--]-->`) // close
-        break
-      default:
-        if (shapeFlag & ShapeFlags.ELEMENT) {
-          renderElementVNode(push, vnode, parentComponent)
-        } else if (shapeFlag & ShapeFlags.COMPONENT) {
-          push(renderComponentVNode(vnode, parentComponent))
-        } else if (shapeFlag & ShapeFlags.TELEPORT) {
-          renderTeleportVNode(push, vnode, parentComponent)
-        } else if (shapeFlag & ShapeFlags.SUSPENSE) {
-          renderVNode(
-            push,
-            normalizeSuspenseChildren(vnode).content,
-            parentComponent
-          )
-        } else {
-          warn(
-            '[@vue/server-renderer] Invalid VNode type:',
-            type,
-            `(${typeof type})`
-          )
-        }
-    }
-  }
-
-  function renderVNodeChildren(
-    push: PushFn,
-    children: VNodeArrayChildren,
-    parentComponent: ComponentInternalInstance
-  ) {
-    for (let i = 0; i < children.length; i++) {
-      renderVNode(push, normalizeVNode(children[i]), parentComponent)
-    }
-  }
-
-  function renderElementVNode(
-    push: PushFn,
-    vnode: VNode,
-    parentComponent: ComponentInternalInstance
-  ) {
-    const tag = vnode.type as string
-    let { props, children, shapeFlag, scopeId, dirs } = vnode
-    let openTag = `<${tag}`
-
-    if (dirs) {
-      props = applySSRDirectives(vnode, props, dirs)
-    }
-
-    if (props) {
-      openTag += ssrRenderAttrs(props, tag)
-    }
-
-    if (scopeId) {
-      openTag += ` ${scopeId}`
-      const treeOwnerId = parentComponent && parentComponent.type.__scopeId
-      // vnode's own scopeId and the current rendering component's scopeId is
-      // different - this is a slot content node.
-      if (treeOwnerId && treeOwnerId !== scopeId) {
-        openTag += ` ${treeOwnerId}-s`
       }
     }
+    push(`</${tag}>`)
+  }
+}
 
-    push(openTag + `>`)
-    if (!isVoidTag(tag)) {
-      let hasChildrenOverride = false
-      if (props) {
-        if (props.innerHTML) {
-          hasChildrenOverride = true
-          push(props.innerHTML)
-        } else if (props.textContent) {
-          hasChildrenOverride = true
-          push(escapeHtml(props.textContent))
-        } else if (tag === 'textarea' && props.value) {
-          hasChildrenOverride = true
-          push(escapeHtml(props.value))
-        }
-      }
-      if (!hasChildrenOverride) {
-        if (shapeFlag & ShapeFlags.TEXT_CHILDREN) {
-          push(escapeHtml(children as string))
-        } else if (shapeFlag & ShapeFlags.ARRAY_CHILDREN) {
-          renderVNodeChildren(
-            push,
-            children as VNodeArrayChildren,
-            parentComponent
-          )
-        }
-      }
-      push(`</${tag}>`)
+function applySSRDirectives(
+  vnode: VNode,
+  rawProps: VNodeProps | null,
+  dirs: DirectiveBinding[]
+): VNodeProps {
+  const toMerge: VNodeProps[] = []
+  for (let i = 0; i < dirs.length; i++) {
+    const binding = dirs[i]
+    const {
+      dir: { getSSRProps }
+    } = binding
+    if (getSSRProps) {
+      const props = getSSRProps(binding, vnode)
+      if (props) toMerge.push(props)
     }
   }
+  return mergeProps(rawProps || {}, ...toMerge)
+}
 
-  function applySSRDirectives(
-    vnode: VNode,
-    rawProps: VNodeProps | null,
-    dirs: DirectiveBinding[]
-  ): VNodeProps {
-    const toMerge: VNodeProps[] = []
-    for (let i = 0; i < dirs.length; i++) {
-      const binding = dirs[i]
-      const {
-        dir: { getSSRProps }
-      } = binding
-      if (getSSRProps) {
-        const props = getSSRProps(binding, vnode)
-        if (props) toMerge.push(props)
-      }
-    }
-    return mergeProps(rawProps || {}, ...toMerge)
+function renderTeleportVNode(
+  push: PushFn,
+  vnode: VNode,
+  parentComponent: ComponentInternalInstance
+) {
+  const target = vnode.props && vnode.props.to
+  const disabled = vnode.props && vnode.props.disabled
+  if (!target) {
+    warn(`[@vue/server-renderer] Teleport is missing target prop.`)
+    return []
   }
-
-  function renderTeleportVNode(
-    push: PushFn,
-    vnode: VNode,
-    parentComponent: ComponentInternalInstance
-  ) {
-    const target = vnode.props && vnode.props.to
-    const disabled = vnode.props && vnode.props.disabled
-    if (!target) {
-      warn(`[@vue/server-renderer] Teleport is missing target prop.`)
-      return []
-    }
-    if (!isString(target)) {
-      warn(
-        `[@vue/server-renderer] Teleport target must be a query selector string.`
-      )
-      return []
-    }
-    renderTeleport(
-      push,
-      push => {
-        renderVNodeChildren(
-          push,
-          vnode.children as VNodeArrayChildren,
-          parentComponent
-        )
-      },
-      target,
-      disabled || disabled === '',
-      parentComponent
+  if (!isString(target)) {
+    warn(
+      `[@vue/server-renderer] Teleport target must be a query selector string.`
     )
+    return []
+  }
+  renderTeleport(
+    push,
+    push => {
+      renderVNodeChildren(
+        push,
+        vnode.children as VNodeArrayChildren,
+        parentComponent
+      )
+    },
+    target,
+    disabled || disabled === '',
+    parentComponent
+  )
+}
+
+export function renderTeleport(
+  parentPush: PushFn,
+  contentRenderFn: (push: PushFn) => void,
+  target: string,
+  disabled: boolean,
+  parentComponent: ComponentInternalInstance
+) {
+  parentPush('<!--teleport start-->')
+
+  let teleportContent: SSRBufferItem
+
+  if (disabled) {
+    contentRenderFn(parentPush)
+    teleportContent = `<!---->`
+  } else {
+    const { getBuffer, push } = createBuffer()
+    contentRenderFn(push)
+    push(`<!---->`) // teleport end anchor
+    teleportContent = getBuffer()
   }
 
-  function renderTeleport(
-    parentPush: PushFn,
-    contentRenderFn: (push: PushFn) => void,
-    target: string,
-    disabled: boolean,
-    parentComponent: ComponentInternalInstance
-  ) {
-    parentPush('<!--teleport start-->')
+  const context = parentComponent.appContext.provides[
+    ssrContextKey as any
+  ] as SSRContext
+  const teleportBuffers =
+    context.__teleportBuffers || (context.__teleportBuffers = {})
+  if (teleportBuffers[target]) {
+    teleportBuffers[target].push(teleportContent)
+  } else {
+    teleportBuffers[target] = [teleportContent]
+  }
 
-    let teleportContent: SSRBufferItem
+  parentPush('<!--teleport end-->')
+}
 
-    if (disabled) {
-      contentRenderFn(parentPush)
-      teleportContent = `<!---->`
+export function renderSlot(
+  slots: Slots | SSRSlots,
+  slotName: string,
+  slotProps: Props,
+  fallbackRenderFn: (() => void) | null,
+  push: PushFn,
+  parentComponent: ComponentInternalInstance
+) {
+  // template-compiled slots are always rendered as fragments
+  push(`<!--[-->`)
+  const slotFn = slots[slotName]
+  if (slotFn) {
+    if (slotFn.length > 1) {
+      // only ssr-optimized slot fns accept more than 1 arguments
+      const scopeId = parentComponent && parentComponent.type.__scopeId
+      slotFn(slotProps, push, parentComponent, scopeId ? ` ${scopeId}-s` : ``)
     } else {
-      const { getBuffer, push } = createBuffer()
-      contentRenderFn(push)
-      push(`<!---->`) // teleport end anchor
-      teleportContent = getBuffer()
+      // normal slot
+      renderVNodeChildren(push, (slotFn as Slot)(slotProps), parentComponent)
     }
-
-    const context = parentComponent.appContext.provides[
-      ssrContextKey as any
-    ] as SSRContext
-    const teleportBuffers =
-      context.__teleportBuffers || (context.__teleportBuffers = {})
-    if (teleportBuffers[target]) {
-      teleportBuffers[target].push(teleportContent)
-    } else {
-      teleportBuffers[target] = [teleportContent]
-    }
-
-    parentPush('<!--teleport end-->')
+  } else if (fallbackRenderFn) {
+    fallbackRenderFn()
   }
-
-  function renderSlot(
-    slots: Slots | SSRSlots,
-    slotName: string,
-    slotProps: Props,
-    fallbackRenderFn: (() => void) | null,
-    push: PushFn,
-    parentComponent: ComponentInternalInstance
-  ) {
-    // template-compiled slots are always rendered as fragments
-    push(`<!--[-->`)
-    const slotFn = slots[slotName]
-    if (slotFn) {
-      if (slotFn.length > 1) {
-        // only ssr-optimized slot fns accept more than 1 arguments
-        const scopeId = parentComponent && parentComponent.type.__scopeId
-        slotFn(slotProps, push, parentComponent, scopeId ? ` ${scopeId}-s` : ``)
-      } else {
-        // normal slot
-        renderVNodeChildren(push, (slotFn as Slot)(slotProps), parentComponent)
-      }
-    } else if (fallbackRenderFn) {
-      fallbackRenderFn()
-    }
-    push(`<!--]-->`)
-  }
-
-  return {
-    renderComponent,
-    renderComponentVNode,
-    renderSlot,
-    renderTeleport
-  }
+  push(`<!--]-->`)
 }
